move hook for superuser auth from app settings to app manager settings
