+++ app/assets/locales/messages_ccodk_default.txt
@@ -732,6 +732,8 @@ settings.server.title=CommCare > Server Settings
 settings.developer.options=Developer Options
 settings.main.title=CommCare > Application Preferences
 settings.advanced.title=CommCare > Advanced
+settings.app.manager.title=App Manager > Settings
+settings.app.manager.option=App Manager Settings
 clear.user.data=Clear User Data
 clear.user.data.warning.title=Clear User Data
 clear.user.data.warning.message=Are you sure you want to clear all mobile user data? All unsynced data will be lost!
