+++ app/res/xml/app_manager_preferences.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+    <Preference android:key="authenticate-as-superuser"/>
+</PreferenceScreen>
